refactor(code): separated constant routes in libs
