--- src/app/(dashboard)/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import LoginPage from "../auth/login/_components/LoginPage";
-
-export default function Home() {
-  return <><LoginPage/></>;
-}
