+++ src/app/_components/sidebar-routes.tsx
@@ -1,68 +1,9 @@
 "use client";
 
-import {
-  BarChart,
-  Compass,
-  Layout,
-  List,
-  MessageCircleQuestion,
-  MessagesSquare,
-  LogOut,
-  Code,
-} from "lucide-react";
-
 import { usePathname } from "next/navigation";
 
 import { SidebarItem } from "./sidebar-item";
-
-const guestRoutes = [
-  {
-    icon: Layout,
-    label: "Dashboard",
-    href: "/",
-  },
-  {
-    icon: Compass,
-    label: "Browse",
-    href: "/search",
-  },
-  {
-    icon: Code,
-    label: "Playground",
-    href: "/playground",
-  },
-  {
-    icon: MessagesSquare,
-    label: "Discussion",
-    href: "/discussion",
-  },
-  {
-    icon: MessageCircleQuestion,
-    label: "Help",
-    href: "/help",
-  },
-];
-
-const teacherRoutes = [
-  {
-    icon: List,
-    label: "Courses",
-    href: "/teacher/courses",
-  },
-  {
-    icon: BarChart,
-    label: "Analytics",
-    href: "/teacher/analytics",
-  },
-];
-
-const commonRoutes = [
-  {
-    icon: LogOut,
-    label: "Logout",
-    href: "/",
-  },
-];
+import { commonRoutes, guestRoutes, teacherRoutes } from "@/lib/sidebar-routes";
 
 export const SidebarRoutes = () => {
   const pathname = usePathname();
