+++ src/app/_components/student-navbar-routes.tsx
@@ -18,42 +18,7 @@ import {
 import { usePathname } from "next/navigation";
 
 import { SidebarItem } from "./sidebar-item";
-
-const guestRoutes = [
-  {
-    icon: CircleUser,
-    label: "Profile",
-    href: "/profile",
-  },
-  {
-    icon: Search,
-    label: "Search",
-    href: "/",
-  },
-  {
-    icon: Slack,
-    label: "Slack",
-    href: "/slack",
-  },
-  {
-    icon: Bell,
-    label: "Notifications",
-    href: "/notiication",
-  },
-];
-
-const teacherRoutes = [
-  {
-    icon: List,
-    label: "Courses",
-    href: "/teacher/courses",
-  },
-  {
-    icon: BarChart,
-    label: "Analytics",
-    href: "/teacher/analytics",
-  },
-];
+import { guestRoutes, teacherRoutes } from "@/lib/navbar-routes";
 
 // const commonRoutes = [
 //   {
