+++ src/app/page.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import LoginPage from "./auth/login/_components/LoginPage";
+
+export default function Home() {
+  return (
+    <>
+      <LoginPage />
+    </>
+  );
+}
